Refactor rendering and UI components for improved performance and functionality
- Optimized GraphicsPipelineStateObject to use arrays and bitsets for resource management.
- Updated Material class to replace constant buffers with property blocks for better memory handling.
- Enhanced RenderPipeline with frame resource management and cleanup methods.
- Integrated VolumetricFog feature into Renderer and adjusted rendering tasks accordingly.
- Added clear color and depth properties to RenderTexture initialization.
- Improved CheckBox UI element with a method to set checked state programmatically.
- Introduced Composite UI element for better handling of vector fields and properties.
- Added error handling in SplitView for child count validation.
- Ensured proper cleanup of property observers in UIElement destructor.
- Fixed interaction zone index calculation in UIFramework for consistency.

+++ Engine/Src/UI/UIFramework.cpp
@@ -184,7 +184,7 @@ namespace Ailu::UI
         }
         else
         {
-            index = _interaction_zones.size();
+            index = (u32)_interaction_zones.size();
             _interaction_zones.push_back({rect, 0});
         }
